Refactor error handling to use uint64 for error IDs and improve stream connection handling with reflection

+++ cgotest/c_tests/bidi_stream_test.c
@@ -1,4 +1,5 @@
 #include <assert.h>
+#include <inttypes.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -10,7 +11,7 @@
 
 typedef struct {
     int done;
-    int done_error_id;
+    uint64_t done_error_id;
     int count;
     char results[16][64];
 } stream_state;
@@ -57,7 +58,7 @@ static void on_read_bytes(uint64_t call_id, void *resp_ptr, int resp_len, FreeFu
     if (resp_free) resp_free(resp_ptr);
 }
 
-static void on_done(uint64_t call_id, int error_id)
+static void on_done(uint64_t call_id, uint64_t error_id)
 {
     stream_state *st = get_state(call_id);
     st->done = 1;
@@ -132,7 +133,7 @@ int main(void) {
 
         wait_done(&st);
         if (!st.done || st.done_error_id != 0) {
-            fprintf(stderr, "expected done with error=0, got done=%d err=%d\n", st.done, st.done_error_id);
+            fprintf(stderr, "expected done with error=0, got done=%d err=%" PRIu64 "\n", st.done, st.done_error_id);
             return 1;
         }
         if (st.count != 3) {
@@ -179,7 +180,7 @@ int main(void) {
 
         wait_done(&st);
         if (!st.done || st.done_error_id != 0) {
-            fprintf(stderr, "expected done with error=0, got done=%d err=%d\n", st.done, st.done_error_id);
+            fprintf(stderr, "expected done with error=0, got done=%d err=%" PRIu64 "\n", st.done, st.done_error_id);
             return 1;
         }
         if (st.count != 3) {
